Start moving some agg and reductions methods
