+++ python/cudf/cudf/tests/groupby/conftest.py
@@ -5,3 +5,10 @@ import pytest
 @pytest.fixture(params=[True, False])
 def as_index(request):
     return request.param
+
+
+@pytest.fixture(
+    params=["min", "max", "idxmin", "idxmax", "count", "sum", "prod", "mean"]
+)
+def groupby_reduction_methods(request):
+    return request.param
